Fix compilation warnings.
Former-commit-id: 09930e9eb8c005fe6835d2bea2e3da5ab9a25cda

+++ server/src/socket/loop.c
@@ -451,7 +451,7 @@ void doeric_server()
 		{
 			if (init_sockets[i].status > Ns_Wait)
 			{
-				if (init_sockets[i].keepalive++ >= SOCKET_KEEPALIVE_TIMEOUT * (1000000 / max_time))
+				if (init_sockets[i].keepalive++ >= (uint32) SOCKET_KEEPALIVE_TIMEOUT * (1000000 / max_time))
 				{
 					FREE_SOCKET(i);
 					continue;
@@ -474,7 +474,7 @@ void doeric_server()
 			pl->socket.status = Ns_Dead;
 		}
 
-		if (pl->socket.status != Ns_Dead && pl->socket.socket_version >= 1052 && pl->socket.keepalive++ >= SOCKET_KEEPALIVE_TIMEOUT * (1000000 / max_time))
+		if (pl->socket.status != Ns_Dead && pl->socket.socket_version >= 1052 && pl->socket.keepalive++ >= (uint32) SOCKET_KEEPALIVE_TIMEOUT * (1000000 / max_time))
 		{
 			pl->socket.status = Ns_Dead;
 		}
